requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pillow
pandas
numpy
